Fix CID 712419, 712420, 712421, 712422: Unintentional sign extension to 64 bits.

+++ src/add-ons/kernel/file_systems/bfs/BlockAllocator.cpp
@@ -1990,8 +1990,9 @@ BlockAllocator::_CheckInodeBlocks(Inode* inode, const char* name)
 			if (status != B_OK)
 				return status;
 
-			int32 maxIndex = (indirect.Length() << fVolume->BlockShift())
-				/ sizeof(block_run);
+			int32 maxIndex =
+				((uint32)indirect.Length() << fVolume->BlockShift())
+					/ sizeof(block_run);
 
 			for (int32 index = 0; index < maxIndex; ) {
 				block_run* runs = (block_run*)cachedDirect.SetTo(
